services/resume-parser-service/requirements.txt
Flask==2.3.2
pdfminer.six==20221105